Desktop/hackai/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
motor==3.3.2
python-dotenv==1.0.1
